Added some informations to the README and remove some warnings when compiling

+++ config.hpp
@@ -63,4 +63,4 @@
     static const bool ctrl_s_save = true;
 
     // The width of a tab '\t' in spaces
-    static uint8_t tab_width = 4;
+    static int tab_width = 4;
